+++ src/reader/lookup/pitch-meta.ts
@@ -1,1 +1,1 @@
-export { composeCompoundPitchPatternFromMeta, localPitchPatternFromMeta, localPitchPatternsFromMeta } from './pitch-meta-pattern';
+export { composeCompoundPitchPatternFromMeta, localPitchPatternFromMeta, localPitchPatternsFromMeta, localPitchPatternsFromMetaLookup } from './pitch-meta-pattern';